Fix format warnings in tviced/state_analyzer.c
state_analyzer.c assumes subtracting two char*s will result in an
unsigned int, which can cause warnings. Lacking a portable format
specifier, just use %ld and cast to long.

Reviewed-on: http://gerrit.openafs.org/681
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/tviced/state_analyzer.c
@@ -737,8 +737,8 @@ print_cb_help(void)
     do { \
         char * _p = (char *)addr; \
         char * _m = (char *)map; \
-        printf("loading structure from address %"AFS_PTR_FMT" (offset %u)\n", \
-               addr, _p-_m); \
+        printf("loading structure from address %"AFS_PTR_FMT" (offset %ld)\n", \
+               addr, (long)(_p-_m)); \
     } while (0)
 
 /* structs */
